MultiModuleProject/requirements.txt
Django==5.2.4
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
pyjokes==0.7.3
pyautogui==0.9.54
gunicorn==21.2.0
whitenoise==6.6.0